requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.3
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
docker==7.0.0
et-xmlfile==1.1.0
Flask==2.2.3
fonttools==4.49.0
frozenlist==1.4.1
future==0.18.3
humanfriendly==10.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.3
mdurl==0.1.2
meraki==1.46.0
multidict==6.0.5
numpy==1.25.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.0
pillow==10.2.0
Pygments==2.17.2
PyJWT==2.6.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
requests-toolbelt==0.10.1
rich==13.7.1
six==1.16.0
tzdata==2023.3
urllib3==2.2.1
webex-bot==0.4.0
webexteamsbot==0.1.4.2
webexteamssdk==1.6.1
websockets==10.2
Werkzeug==2.2.3
XlsxWriter==3.2.0
yarl==1.9.4
